refactor(stepper/nav-drawer): optimize CSS
