+++ src/components/nav-drawer/themes/header-item.ts
@@ -2,38 +2,36 @@ import { css } from 'lit';
 
 // Shared Styles
 import { styles as bootstrap } from './shared/header-item/header-item.bootstrap.css.js';
-import { styles as fluent } from './shared/header-item/header-item.fluent.css.js';
-import { styles as indigo } from './shared/header-item/header-item.indigo.css.js';
-import { styles as material } from './shared/header-item/header-item.material.css.js';
+import { styles as shared } from './shared/header-item/header-item.common.css.js';
 import { Themes } from '../../../theming/types.js';
 
 const light = {
   bootstrap: css`
-    ${bootstrap}
+    ${shared} ${bootstrap}
   `,
   material: css`
-    ${material}
+    ${shared}
   `,
   fluent: css`
-    ${fluent}
+    ${shared}
   `,
   indigo: css`
-    ${indigo}
+    ${shared}
   `,
 };
 
 const dark = {
   bootstrap: css`
-    ${bootstrap}
+    ${shared} ${bootstrap}
   `,
   material: css`
-    ${material}
+    ${shared}
   `,
   fluent: css`
-    ${fluent}
+    ${shared}
   `,
   indigo: css`
-    ${indigo}
+    ${shared}
   `,
 };
