+++ src/components/nav-drawer/themes/item.ts
@@ -1,39 +1,38 @@
 import { css } from 'lit';
 
 // Shared Styles
-import { styles as bootstrap } from './shared/item/item.bootstrap.css.js';
+import { styles as shared } from './shared/item/item.common.css.js';
 import { styles as fluent } from './shared/item/item.fluent.css.js';
 import { styles as indigo } from './shared/item/item.indigo.css.js';
-import { styles as material } from './shared/item/item.material.css.js';
 import { Themes } from '../../../theming/types.js';
 
 const light = {
   bootstrap: css`
-    ${bootstrap}
+    ${shared}
   `,
   material: css`
-    ${material}
+    ${shared}
   `,
   fluent: css`
-    ${fluent}
+    ${shared} ${fluent}
   `,
   indigo: css`
-    ${indigo}
+    ${shared} ${indigo}
   `,
 };
 
 const dark = {
   bootstrap: css`
-    ${bootstrap}
+    ${shared}
   `,
   material: css`
-    ${material}
+    ${shared}
   `,
   fluent: css`
-    ${fluent}
+    ${shared} ${fluent}
   `,
   indigo: css`
-    ${indigo}
+    ${shared} ${indigo}
   `,
 };
