+++ src/components/nav-drawer/themes/shared/container/nav-drawer.bootstrap.scss
@@ -2,54 +2,9 @@
 @use '../../light/themes' as *;
 
 $theme: $bootstrap;
-$background: var-get($theme, 'background') !default;
 
 :host {
-    [part='base'] {
-        border-radius: var-get($theme, 'border-radius');
-        background: $background;
-        border: none;
-        box-shadow: none;
-    }
-
     [part='overlay'] {
         background: var-get($overlay-bootstrap, 'background-color');
-    }
-}
-
-[part='mini'] {
-    background: $background;
-}
-
-:host([position='start']) [part='mini'],
-:host([position='end']) [part='mini'] {
-    box-shadow: var-get($theme, 'elevation');
-}
-
-:host([position='start']) [part='mini'] {
-    border-inline-end-color: var-get($theme, 'border-color');
-}
-
-:host([position='end']) [part='mini'] {
-    border-inline-start-color: var-get($theme, 'border-color');
-}
-
-:host([open][position='end']) {
-    [part='base'] {
-        border-inline-start-color: var-get($theme, 'border-color');
-    }
-}
-
-:host([position='relative']) {
-    [part='base'],
-    [part~='mini'] {
-        border-inline-end-color: var-get($theme, 'border-color');
-    }
-}
-
-:host([open]:not([position='relative'])) {
-    [part='base'] {
-        border-inline-end-color: var-get($theme, 'border-color');
-        box-shadow: var-get($theme, 'elevation');
     }
 }
