+++ src/components/nav-drawer/themes/shared/header-item/header-item.bootstrap.scss
@@ -1,10 +1,5 @@
 @use 'styles/utilities' as *;
-@use '../../light/themes' as *;
-
-$theme: $bootstrap;
 
 :host {
     @include type-style('caption');
-
-    color: var-get($theme, 'item-header-text-color');
 }
