+++ src/components/nav-drawer/themes/shared/header-item/header-item.common.scss
