--- src/components/nav-drawer/themes/shared/header-item/header-item.indigo.scss
@@ -1,10 +1,0 @@
-@use 'styles/utilities' as *;
-@use '../../light/themes' as *;
-
-$theme: $indigo;
-
-:host {
-    @include type-style('subtitle-1');
-
-    color: var-get($theme, 'item-header-text-color');
-}
