--- src/components/nav-drawer/themes/shared/item/item.bootstrap.scss
@@ -1,38 +1,0 @@
-@use 'styles/utilities' as *;
-@use '../../light/themes' as *;
-
-$theme: $bootstrap;
-
-[part~='base'] {
-    border-radius: var-get($theme, 'item-border-radius');
-    color: var-get($theme, 'item-text-color');
-
-    &:hover,
-    &:focus-within {
-        background: var-get($theme, 'item-hover-background');
-        color: var-get($theme, 'item-hover-text-color');
-
-        ::slotted(igc-icon) {
-            color: var-get($theme, 'item-hover-icon-color');
-        }
-    }
-
-    ::slotted(igc-icon) {
-        color: var-get($theme, 'item-icon-color');
-    }
-}
-
-[part='content'] {
-    @include type-style('subtitle-2');
-}
-
-:host([active]) {
-    [part~='base'] {
-        background: var-get($theme, 'item-active-background');
-        color: var-get($theme, 'item-active-text-color');
-
-        ::slotted(igc-icon) {
-            color: var-get($theme, 'item-active-icon-color');
-        }
-    }
-}
