+++ src/components/stepper/themes/step/shared/step.common.scss
