+++ src/components/stepper/themes/step/shared/step.indigo.scss
@@ -3,161 +3,24 @@
 
 $theme: $indigo;
 
-[part~='title'] {
-    color: var-get($theme, 'title-color');
-}
-
-[part~='subtitle'] {
-    color: var-get($theme, 'subtitle-color');
-}
-
-[part~='header'] {
-    background: var-get($theme, 'step-background');
-    border-radius: var-get($theme, 'border-radius-step-header');
-
-    &::before,
-    &::after {
-        border-block-start-style: var-get($theme, 'step-separator-style');
-        border-block-start-color: var-get($theme, 'step-separator-color');
-    }
-
-    &:hover {
-        background: var-get($theme, 'step-hover-background');
-        color: var-get($theme, 'title-hover-color')
-    }
-
-    &:focus,
-    &:focus-within,
-    &:active {
-        background: var-get($theme, 'step-focus-background');
-        color: var-get($theme, 'title-focus-color');
-    }
-}
-
-[part~='header-container'],
-[part~='header'] {
-    &::before {
-        border-block-start-color: var-get($theme, 'step-separator-color');
-        border-block-start-style: var-get($theme, 'step-separator-style');
-    }
-}
-
-[part~='body'] {
-    &::before {
-        border-inline-start-color: var-get($theme, 'step-separator-color');
-    }
-}
-
 [part~='indicator'] {
-    border-radius: var-get($theme, 'border-radius-indicator');
-    background: var-get($theme, 'indicator-background');
-    color: var-get($theme, 'indicator-color');
     box-shadow: inset 0 0 0 var(--indicator-box-shadow-size) var-get($theme, 'indicator-outline');
 }
 
 [part~='complete-start'] {
-    &::after {
-        border-block-start-style: var-get($theme, 'complete-step-separator-style');
-        border-block-start-color: var-get($theme, 'complete-step-separator-color');
-    }
-
-    [part~='header'] {
-        &::after {
-            border-block-start-style: var-get($theme, 'complete-step-separator-style');
-            border-block-start-color: var-get($theme, 'complete-step-separator-color');
-        }
-    }
-
     [part~='indicator'] {
-        background: var-get($theme, 'complete-indicator-background');
-        color: var-get($theme, 'complete-indicator-color');
         box-shadow: inset 0 0 0 var(--indicator-box-shadow-size) var-get($theme, 'complete-indicator-outline');
-    }
-
-    [part~='title'] {
-        color: var-get($theme, 'complete-title-color');
-    }
-
-    [part~='subtitle'] {
-        color: var-get($theme, 'complete-subtitle-color');
-    }
-}
-
-:host([complete]) {
-    [part~='body'] {
-        &::before {
-            border-inline-start-style: var-get($theme, 'complete-step-separator-style');
-            border-inline-start-color: var-get($theme, 'complete-step-separator-color');
-        }
-    }
-}
-
-[part~='complete-end'] {
-    &::before {
-        border-block-start-style: var-get($theme, 'complete-step-separator-style');
-        border-block-start-color: var-get($theme, 'complete-step-separator-color');
-    }
-
-    [part~='header'] {
-        &::before {
-            border-block-start-style: var-get($theme, 'complete-step-separator-style');
-            border-block-start-color: var-get($theme, 'complete-step-separator-color');
-        }
-    }
-}
-
-[part~='invalid'] {
-    [part~='indicator'] {
-        color: var-get($theme, 'invalid-indicator-color');
-        background: var-get($theme, 'invalid-indicator-background');
-        box-shadow: inset 0 0 0 var(--indicator-box-shadow-size) var-get($theme, 'invalid-indicator-outline');
-    }
-
-    [part~='title'],
-    [part~='subtitle'] {
-        color: var-get($theme, 'invalid-title-color');
     }
 }
 
 :host([active]) {
     [part~='indicator'] {
-        background: var-get($theme, 'current-indicator-background');
-        color: var-get($theme, 'current-indicator-color');
         box-shadow: inset 0 0 0 var(--indicator-box-shadow-size) var-get($theme, 'current-indicator-outline');
-    }
-
-    [part~='title'] {
-        color: var-get($theme, 'current-title-color');
-        font-weight: 600;
-    }
-
-    [part~='subtitle'] {
-        color: var-get($theme, 'current-subtitle-color');
     }
 }
 
 [part~='disabled'] {
-    &:focus,
-    &:focus-visible,
-    &:focus-within,
-    &:active,
-    &:hover {
-        [part='header'] {
-            background: var-get($theme, 'step-background');
-        }
-    }
-
     [part~='indicator'] {
-        color: var-get($theme, 'disabled-indicator-color');
-        background: var-get($theme, 'disabled-indicator-background');
         box-shadow: inset 0 0 0 var(--indicator-box-shadow-size) var-get($theme, 'disabled-indicator-outline');
-    }
-
-    [part~='title'] {
-        color: var-get($theme, 'disabled-title-color');
-    }
-
-    [part~='subtitle'] {
-        color: var-get($theme, 'disabled-subtitle-color');
     }
 }
